requirements.txt
Django==2.2
djangorestframework==3.12.2
psycopg2-binary==2.8.3
requests==2.25.1
gunicorn==19.9.0
python-memcached==1.59
whitenoise==5.2.0
